% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_obtaining_gof.R
\name{gof}
\alias{gof}
\title{Step Three: Obtain Goodness of Fit Figures for a set of computed models}
\usage{
gof(mfits, gof)
}
\arguments{
\item{mfits}{A list with four components:
models: A list with one or two elements. Each element has all possible model fits
in it. Each element represents one distributional parameter
modelids: A character vector with model id's corresponding to the models
in the models element
expl_names: expl_names
npar: Number of parameters
method: One of "lm", "gamlss".}

\item{gof}{Goodness of fit which should be obtained for all models. Which is available
depends on the method with which the models were computed. Currently:
"AIC", "r.squared", "loglik", "deviance" for "lm" method and "AIC", "deviance" for
"gamlss" method.}
}
\value{
A gof object (type list), which has the following structure
}
\description{
Step Three: Obtain Goodness of Fit Figures for a set of computed models
}
