% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_obtaining_gof.R
\name{gof_gamlss}
\alias{gof_gamlss}
\title{Internal: Get GOF's for models computed with "gamlss" method}
\usage{
gof_gamlss(mfits, gof)
}
\description{
Internal: Get GOF's for models computed with "gamlss" method
}
\keyword{internal}
