% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_obtaining_gof.R
\name{gof_getter}
\alias{gof_getter}
\title{Function to obtain specific GOF for all given models}
\usage{
gof_getter(models, gof)
}
\description{
Function to obtain specific GOF for all given models
}
\keyword{internal}
