% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_obtaining_gof.R
\name{gof_lm}
\alias{gof_lm}
\title{Internal: Get GOF's for models computed with "lm" method}
\usage{
gof_lm(mfits, gof)
}
\description{
Internal: Get GOF's for models computed with "lm" method
}
\keyword{internal}
