% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_as_goodfit.R
\name{goodfit}
\alias{goodfit}
\title{Create a goodfit object}
\usage{
goodfit(gofs, model_ids, expl_names, npar = 1, method = "unknown", gof)
}
\arguments{
\item{gofs}{A list with either one or two components, depending on whether
two parameters (e.g. mu and sigma) have to be modeled, or just one (most
cases). The first element is a vector called 'mu' holding the goodness of
fit numbers as a numeric vector. If a second parameter is also modeled, the
second element is called 'sigma'.}

\item{model_ids}{A character vector of the same length as \code{gofs$mu}. Is
used for indexing the \code{gof} vectors, so it has to be in the right
order compared to gofs.}

\item{expl_names}{A character vector with the variable or variable group
names.}

\item{npar}{Number of parameters for which patitioning should be done.
Defaults to 1. Has to have the same length as \code{gofs}}

\item{method}{Method from which the goodnesses of fit were extracted. Not of
importance for partitioning so it defaults to \code{"unknown"}.}

\item{gof}{Name of the goodness of fit measure which was extracted from the
model. Is used for \code{\link{plot_ghp}}.}
}
\value{
A gof object (type list), which has the following elements:
  \enumerate{\item \strong{gofs}: A list with \code{npar} elements, each
  being a vector with the goodness of fits of the models. \item
  \strong{model_ids}: A character vector with id's of the models: e.g. "x0",
  "x1", ... \item \strong{expl_names}: Names of explanatory variables
  (grouped) or names of groups (grouped). \item \strong{npar} Number of
  modeled paramateters. Can be 1 or 2. \item \strong{method}: The method used
  to compute models. Can be one of "lm" or "gamlss". \item \strong{gof}: A
  single character depicting the goodness of fit that was extracted.}
}
\description{
Takes goodness of fit figures and outputs a goodfit object for further
partitioning with \code{\link{part}}.
}
\details{
Use this function if you have models and/or goodness of fit figures which are
not yet implemented in \code{ghp} but still shall be partitioned with
\code{\link{part}}.
}
