% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{india}
\alias{india}
\title{Malnutrition in india}
\format{An object of class \code{data.frame} with 391 rows and 8 columns.}
\usage{
india
}
\description{
Malnutrition in india
}
\keyword{datasets}
