% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_creation.R
\name{mfit}
\alias{mfit}
\title{Step Two: Fit all possible model combinations}
\usage{
mfit(data, method, npar = 1)
}
\arguments{
\item{data}{A list with two components: \code{dep}, a vector with the dependent
variable, and \code{indep}, with the independent variables. The independent
variables can be grouped or not. This list is best created by \code{\link{indep_tf}}.}
}
\value{
A list with the following components:
}
\description{
Step Two: Fit all possible model combinations
}
