% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_creation.R
\name{mfit_gamlss}
\alias{mfit_gamlss}
\title{Fit all possible model combinations for "gamlss" model class}
\usage{
mfit_gamlss(data, npar, combinations)
}
\description{
Fit all possible model combinations for "gamlss" model class
}
\keyword{internal}
