% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_model_creation.R
\name{mfit_lm}
\alias{mfit_lm}
\title{Fit all possible model combinations for "lm" model class}
\usage{
mfit_lm(data, combinations)
}
\description{
Fit all possible model combinations for "lm" model class
}
\keyword{internal}
