% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfitter.R
\name{model_fitter}
\alias{model_fitter}
\title{Internal: Function to fit all model combinations}
\usage{
model_fitter(dep, indep = NULL, combinations = NULL, method, npar = 1,
  empty = FALSE)
}
\description{
Internal: Function to fit all model combinations
}
