% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_partitioning.R
\name{part}
\alias{part}
\title{Step Four: Do hierarchical partitioning}
\usage{
part(gofs_list)
}
\arguments{
\item{gofs_list}{A goodfit object, preferably made with \code{\link{goodfit}}
or \code{\link{gof}}}
}
\description{
Step Four: Do hierarchical partitioning
}
