% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_partitioning.R
\name{part_core}
\alias{part_core}
\title{Core function of part(). Does the partitioning.}
\usage{
part_core(gofs_vector, expl_names, model_ids)
}
\description{
Core function of part(). Does the partitioning.
}
\keyword{internal}
