% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{perm_cpp}
\alias{perm_cpp}
\title{Core partitioning function}
\usage{
perm_cpp(perms, model_ids, gofs_vector)
}
\arguments{
\item{model_ids}{A numeric vector with unique model id's}

\item{gofs_vector}{A numeric vector with goodness-of-fit numbers}

\item{perm}{All permutations in a matrix}
}
\description{
Computes the corresponding difference of a model combination with or without a variable (first one in the permutation). Written in C++.
}
