% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_partitioning.R
\name{perm_diff}
\alias{perm_diff}
\title{Get differences of GOF's for a given combination of covariate numbers}
\usage{
perm_diff(perm, model_ids, gofs_vector)
}
\description{
Get differences of GOF's for a given combination of covariate numbers
}
\keyword{internal}
