% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfitter.R
\name{selector}
\alias{selector}
\title{Internal: Function to make a dataframe out of selected list parts}
\usage{
selector(list, selection)
}
\description{
Internal: Function to make a dataframe out of selected list parts
}
