% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_obtaining_gof.R
\name{gof}
\alias{gof}
\title{Obtain Goodness of Fit Figures}
\usage{
gof(mfits, gof)
}
\arguments{
\item{mfits}{A list with five components: \enumerate{ \item \strong{models}:
A list with one or two elements. Each element has all possible model fits
in it. Each element represents one distributional parameter. \item
\strong{modelids}: A character vector with model id's corresponding to the
models. in the models element \item \strong{expl_names}: names of
explanatory variables (ungrouped case) or group names (grouped case). \item
\strong{npar}: Number of parameters which are modeled. \item
\strong{method}: One of "lm", "gamlss". } Best created by
\code{\link{mfit}}.}

\item{gof}{Goodness of fit which should be obtained for all models. Which is
available depends on the method with which the models were computed.
Currently: \itemize{\item \strong{lm method}: "AIC", "r.squared", "loglik",
"deviance" \item \strong{gamlss method}: "AIC", "deviance", "loglik", "R2e}}
}
\value{
A gof object (type list), which has the following elements:
  \enumerate{\item \strong{gofs}: A list with \code{npar} elements, each
  being a vector with the goodness of fits of the models. \item
  \strong{model_ids}: A character vector with id's of the models: e.g. "x0",
  "x1", ... \item \strong{expl_names}: Names of explanatory variables
  (grouped) or names of groups (grouped). \item \strong{npar} Number of
  modeled paramateters. Can be 1 or 2. \item \strong{method}: The method used
  to compute models. Can be one of "lm" or "gamlss". \item \strong{gof}: A
  single character depicting the goodness of fit that was extracted.}
}
\description{
Step Three: Obtain Goodness of Fit Figures for a set of computed models
}
\details{
This function takes a list generated by \code{\link{mfit}} to obtain all
possible models and then exports a 'goodfit' object, used for further
partitioning.
}
