% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_getter.R
\name{gof_getter}
\alias{gof_getter}
\title{Internal function to get GOF of all fitted models (including the intercept model)}
\usage{
gof_getter(m0, models, gof, npar = 1)
}
\description{
Internal function to get GOF of all fitted models (including the intercept model)
}
\keyword{internal}
