% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_transform.R
\name{indep_tf}
\alias{indep_tf}
\title{Internal: Transform the independent variables for obtaining goodness of fit}
\usage{
indep_tf(depname, data, group_df = NULL)
}
\description{
Internal: Transform the independent variables for obtaining goodness of fit
}
