% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.R
\name{perm_diff}
\alias{perm_diff}
\title{Get differences of GOF's for a given combination of covariate numbers}
\usage{
perm_diff(perm, names, gofs)
}
\description{
Get differences of GOF's for a given combination of covariate numbers
}
\keyword{internal}
